idoit-feature (#39)
* - fix bug

* - fix the idiot role error

* - fix idiot e2e tests

* Idiot Feature
- fix the integration tests
- fix sheriff voting weight

* - lint

* - lint

* fix: remove execSync calls from E2E tests that fail in CI environment

- Remove execSync import from child_process
- Remove local script calls that don't work in CI
- Keep other debugging information for test troubleshooting

+++ backend/src/main/kotlin/com/werewolf/game/DomainEvent.kt
@@ -23,7 +23,7 @@ sealed class DomainEvent {
     @JsonTypeName("VoteSubmitted")
     data class VoteSubmitted(val gameId: Int, val voterUserId: String) : DomainEvent()
     @JsonTypeName("VoteTally")
-    data class VoteTally(val gameId: Int, val eliminatedUserId: String?, val tally: Map<String, Int>) : DomainEvent()
+    data class VoteTally(val gameId: Int, val eliminatedUserId: String?, val tally: Map<String, Double>) : DomainEvent()
     @JsonTypeName("PlayerEliminated")
     data class PlayerEliminated(val gameId: Int, val userId: String, val role: PlayerRole) : DomainEvent()
     @JsonTypeName("HunterShot")
